feat: refactor schedule
