+++ src/components/Block.astro
@@ -1,5 +1,5 @@
 ---
-import type { Block } from "../utils/schedule";
+import type { Block } from "../utils/schedule/types";
 import Accordion from "./Accordion.vue";
 import RenderTrack from "./RenderTrack.astro";
 import ScheduleEntryCard from "./ScheduleEntryCard/ScheduleEntryCard.astro";
