+++ src/components/RenderTrack.astro
@@ -2,8 +2,8 @@
 import {
   getShortestTalkDuration,
   getSlotsAvaliable,
-  type Block,
-} from "../utils/schedule";
+} from "../utils/schedule/functions";
+import type { Block } from "../utils/schedule/types";
 import Track from "./Track.astro";
 
 interface Props {
