+++ src/pages/schedule.astro
@@ -2,7 +2,7 @@
 import Block from "../components/Block.astro";
 import Toggle from "../components/Toggle.astro";
 import ScheduleLayout from "../layouts/ScheduleLayout.astro";
-import { schedule } from "../utils/schedule";
+import { schedule } from "../utils/schedule/schedule";
 ---
 
 <ScheduleLayout>
