+++ src/pages/talks/[talk].astro
@@ -5,7 +5,7 @@ import {
   getDisciplineColorPair,
   getDisciplineName,
 } from "../../utils/discipline";
-import { getTalkById } from "../../utils/schedule";
+import { getTalkById } from "../../utils/schedule/functions";
 
 const exampleDescription =
   "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
